+++ src/main/kotlin/me/odinmod/odin/clickgui/settings/impl/ActionSetting.kt
@@ -20,7 +20,7 @@ class ActionSetting(
 
     private val textWidth by lazy { NVGRenderer.textWidth(name, 16f, NVGRenderer.defaultFont) }
 
-    override fun render(x: Float, y: Float, mouseX: Double, mouseY: Double): Float {
+    override fun render(x: Float, y: Float, mouseX: Float, mouseY: Float): Float {
         super.render(x, y, mouseX, mouseY)
         val height = getHeight()
 
@@ -30,7 +30,7 @@ class ActionSetting(
         return height
     }
 
-    override fun mouseClicked(mouseX: Double, mouseY: Double, mouseButton: Int): Boolean {
+    override fun mouseClicked(mouseX: Float, mouseY: Float, mouseButton: Int): Boolean {
         return if (mouseButton != 0 || !isHovered) false
         else {
             action()
